pyproject.toml
[tool.poetry]
name = "popos-utils"
version = "0.1.0"
description = ""
authors = ["templepmet"]
readme = "README.md"
packages = [{include = "popos_utils"}]

[tool.poetry.dependencies]
python = "^3.10"
python-xlib = "^0.33"


[tool.poetry.group.dev.dependencies]
black = {version = "^23.1a1", allow-prereleases = true}
flake8 = "^6.0.0"
mypy = "^0.991"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
